Add files via upload
I still need to do a lot, most of which is the functionality, but I need to wait till tommorow, cause I just accidentelly stared at my computer screen for about 20 minutes while trying to find something, and I need sleep. I'll finish it tommorow, as most of the functionality should be copy-paste with minor functionality changes. The only problem is tedium.